requirements.txt
# Core Flask App
Flask==2.3.3
flask-socketio==5.3.6
python-dotenv==1.0.1

# Requests & Networking
requests==2.31.0
httpx==0.27.0  # Keep this for OpenAI SDK compatibility
httpcore==1.0.5

# OpenAI & OpenRouter APIs
openai==1.30.1

# Translations
deep_translator==1.11.4  # Preferred over googletrans

# Voice & Audio
SpeechRecognition==3.10.1
PyAudio==0.2.13

# PDF & QR
reportlab==4.1.0
qrcode==7.4.2

# AI Model
scikit-learn==1.3.0
joblib==1.4.0

# Image Processing
opencv-python==4.9.0.80

# Optional: Face detection (if used)
deepface==0.0.79

# Deployment
gunicorn==21.2.0
eventlet==0.35.1
